refactor: inline ToolEditApprovalEvents, delete 75-line ceremony file
The 3 event subscriptions don't need a separate module with interfaces,
factory functions, and shared state objects. Inline them directly in
ProgressEventHandler.setupEventListeners().
